start breaking out a ui feature to set an input step (#253)
* start

* refactoring

* working

* version bump
