+++ examples/index.html
@@ -16,7 +16,6 @@
   <script src="../node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
   <script src="../dist/image-sequencer.js" charset="utf-8"></script>
   <script src="lib/urlHash.js" charset="utf-8"></script>
-  <script src="lib/imageSelection.js" charset="utf-8"></script>
   <script src="lib/defaultHtmlStepUi.js" charset="utf-8"></script>
   <script src="lib/defaultHtmlSequencerUi.js" charset="utf-8"></script>
   <script src="demo.js" charset="utf-8"></script>
