+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "image-sequencer",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "description": "A modular JavaScript image manipulation library modeled on a storyboard.",
   "main": "src/ImageSequencer.js",
   "scripts": {
